+++ src/main/java/kr/bb/wishlist/cart/mapper/CartMapper.java
@@ -5,7 +5,7 @@ import kr.bb.wishlist.cart.entity.CartEntity;
 public class CartMapper {
 
   public static CartEntity getCartEntityWithUpdatedSelectedQuantity(CartEntity cartEntity,
-      int totalQuantity) {
+      Long totalQuantity) {
     return CartEntity.builder().cartCompositekey(cartEntity.getCartCompositekey())
         .selectedQuantity(totalQuantity).build();
   }
